src/emailservice/requirements.txt
setuptools>=68.0.0
# EmailService requirements - Cloud-Agnostic
# Compatible with Python 3.12+
#
# REMOVED GCP packages:
#   - google-cloud-profiler, google-cloud-trace
#   - google-api-core, google-api-python-client
#   - google-auth

# Core gRPC - 1.62+ required for Python 3.12 compatibility
grpcio==1.62.2
grpcio-health-checking==1.62.2

# Templating
jinja2==3.1.4
markupsafe==2.1.5

# Logging
python-json-logger==2.0.7

# OpenTelemetry (optional - works with Jaeger, AWS X-Ray ADOT, etc.)
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation-grpc==0.48b0

# Transitive
protobuf==4.25.3
googleapis-common-protos==1.63.0
backoff==2.2.1
deprecated==1.2.14
wrapt==1.16.0
typing-extensions==4.11.0
requests==2.31.0
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
urllib3==2.2.1
